added possibility to append responses to alrdy existing file instead of having to always override.

+++ getResponsesFromLLM.py
@@ -120,6 +120,11 @@ if __name__ == '__main__':
         else:
             out_path = args.output.strip("/")
             out_path = f"/{out_path}/{modelname}_responses.json"
-
-        with open(out_path, 'w', encoding='utf-8') as write_file:
-            json.dump(json_template, write_file, indent=4, ensure_ascii=False)
+        if os.path.exists(out_path):
+            with open(out_path, 'rw', encoding='utf-8') as write_file:
+                oldFile = dict(json.load(write_file))
+                oldFile.update(json_template)
+                json.dump(json_template, write_file, indent=4, ensure_ascii=False)
+        else:
+            with open(out_path, 'w', encoding='utf-8') as write_file:
+                json.dump(json_template, write_file, indent=4, ensure_ascii=False)
